+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format of this changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+-   removed query params to calls list numbers
+
 <a name="0.4.0"></a>
 
 ## [0.4.0] - 2021-04-07
@@ -13,7 +17,6 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 -   query params to incoming numbers
 -   query params to available numbers
--   query params to calls list numbers
 
 ## [0.3.0] - 2021-04-07